Reworked how the functions in the population work
Now elites are separated from the rest of individuals.
After that selection occurs on the remaining individuals.
Pairing is done to produce the desired amount of children. This way crossover rate can be adjusted.

Remaining individuals are picked from the parent individuals.

And then mutations are performed on that population.

And then only the fittest individuals from children and parents are left in the population.

Finally, the elites are added back into the new generation